Move font loading code into own module
