+++ Cargo.toml
@@ -34,7 +34,7 @@ stable_bst = "0.2.0"
 maplit = "0.1.4"
 multi_mut = "0.1.3"
 
-font-loader = "0.4.2"
+font-loader = "0.5.0"
 
 [dev-dependencies]
 find_folder = "0.3.0"
