+++ src/draw/text.rs
@@ -25,7 +25,7 @@ impl TextState {
     pub fn measure(&self) -> Size {
         let line_height = self.line_height();
         let mut resources = resources();
-        let font = resources.get_font(&self.font);
+        let font = resources.font_loader.get_font(&self.font);
         Size::from_untyped(&text_layout::get_text_size(
             &self.text,
             &font.info,
@@ -42,7 +42,7 @@ impl TextState {
     pub fn text_fits(&self, text: &str, bounds: Rect) -> bool {
         let line_height = self.line_height();
         let mut resources = resources();
-        let font = resources.get_font(&self.font);
+        let font = resources.font_loader.get_font(&self.font);
         let height = text_layout::get_text_height(
             text,
             &font.info,
@@ -55,7 +55,7 @@ impl TextState {
     fn get_line_rects(&self, bounds: Rect) -> Vec<Rect> {
         let line_height = self.line_height();
         let mut resources = resources();
-        let font = resources.get_font(&self.font);
+        let font = resources.font_loader.get_font(&self.font);
         text_layout::get_line_rects(
             &self.text,
             bounds.to_untyped(),
@@ -69,7 +69,7 @@ impl TextState {
         let line_height = self.line_height();
         let descent = self.v_metrics().descent;
         let mut resources = resources();
-        let font = resources.get_font(&self.font);
+        let font = resources.font_loader.get_font(&self.font);
         text_layout::get_positioned_glyphs(
             &self.text,
             bounds.to_untyped(),
@@ -86,11 +86,11 @@ impl TextState {
             }).collect()
     }
     fn font_instance_key(&self) -> FontInstanceKey {
-        *resources().get_font_instance(&self.font, self.font_size)
+        *resources().font_loader.get_font_instance(&self.font, self.font_size)
     }
     fn v_metrics(&self) -> VMetrics {
         let mut resources = resources();
-        let font = resources.get_font(&self.font);
+        let font = resources.font_loader.get_font(&self.font);
         font.info.v_metrics(Scale::uniform(self.font_size))
     }
 }
@@ -102,7 +102,7 @@ impl Draw for TextState {
             let line_rects = self.get_line_rects(bounds);
             let v_metrics = self.v_metrics();
             let mut resources = resources();
-            let font = resources.get_font(&self.font);
+            let font = resources.font_loader.get_font(&self.font);
             for mut rect in line_rects {
                 render::draw_rect_outline(rect, CYAN, renderer);
                 rect.origin.y = rect.bottom() + v_metrics.descent;
