+++ src/resources/mod.rs
@@ -1,5 +1,6 @@
 #[macro_use]
 pub mod id;
+pub mod font;
 
 use std::sync::{Mutex, MutexGuard};
 use std::collections::HashMap;
@@ -7,24 +8,18 @@ use std::default::Default;
 
 use webrender::api::*;
 use image;
-use rusttype;
-use app_units;
-use font_loader::system_fonts;
-
-use text_layout;
 
 use self::id::{Id, IdGen};
+use self::font::FontLoader;
 
 use style::Theme;
-
-pub type Font = rusttype::Font<'static>;
 
 lazy_static! {
     static ref RES: Mutex<Resources> = Mutex::new(Resources::new());
 }
 
-pub fn init_resources(render_api: RenderApi) {
-    RES.try_lock().unwrap().render = Some(render_api);
+pub fn init_resources(render_api: RenderApiSender) {
+    RES.try_lock().unwrap().set_render_api(render_api);
 }
 // Allow global access to Resources
 pub fn resources() -> MutexGuard<'static, Resources> {
@@ -33,11 +28,6 @@ pub fn resources() -> MutexGuard<'static, Resources> {
 
 named_id!(WidgetId);
 
-#[derive(Clone)]
-pub struct FontInfo {
-    pub key: FontKey,
-    pub info: Font,
-}
 
 #[derive(Debug, Copy, Clone)]
 pub struct ImageInfo {
@@ -80,8 +70,7 @@ impl<I: Id, T> Map<I, T> {
 
 pub struct Resources {
     pub render: Option<RenderApi>,
-    pub fonts: HashMap<String, FontInfo>,
-    pub font_instances: HashMap<(String, app_units::Au), FontInstanceKey>,
+    pub font_loader: FontLoader,
     pub images: HashMap<String, ImageInfo>,
     pub texture_descriptors: HashMap<u64, ImageDescriptor>,
     pub widget_id: IdGen<WidgetId>,
@@ -92,8 +81,7 @@ impl Default for Resources {
     fn default() -> Self {
         Resources {
             render: None,
-            fonts: HashMap::new(),
-            font_instances: HashMap::new(),
+            font_loader: FontLoader::new(),
             images: HashMap::new(),
             texture_descriptors: HashMap::new(),
             widget_id: IdGen::new(),
@@ -140,49 +128,9 @@ impl Resources {
         self.images.insert(name.to_owned(), image_info);
         &self.images[name]
     }
-
-    #[deprecated(note = "may panic, instead of this use get_font_or_load_from_system or get_font_if_present")]
-    pub fn get_font(&mut self, name: &str) -> &FontInfo {
-        self.get_font_or_load_from_system(name).unwrap()
-    }
-
-    pub fn get_font_if_present(&mut self, name: &str) -> Option<&FontInfo> {
-        self.fonts.get(name)
-    }
-
-    pub fn get_font_or_load_from_system(&mut self, name: &str) -> Result<&FontInfo, ::std::io::Error> {
-        if !self.fonts.contains_key(name) {
-            return self.add_font(name, try!(load_system_font_by_family_name(name)));
-        }
-        Ok(&self.fonts[name])
-    }
-
-    pub fn add_font(&mut self, name: &str, font_bytes: Vec<u8>) -> Result<&FontInfo, ::std::io::Error> {
-        let font = try!(font_from_bytes(font_bytes.clone()));
-
-        let key = self.render.as_ref().unwrap().generate_font_key();
-        let mut resources = ResourceUpdates::new();
-        resources.add_raw_font(key, font_bytes, 0);
-
-        self.render.as_ref().unwrap().update_resources(resources);
-        let font_info = FontInfo { key: key, info: font };
-        self.fonts.insert(name.to_owned(), font_info);
-
-        Ok(&self.fonts[name])
-    }
-
-    #[cfg_attr(feature = "cargo-clippy", allow(map_entry))]
-    pub fn get_font_instance(&mut self, name: &str, font_size: f32) -> &FontInstanceKey {
-        let font_key = self.get_font(name).key;
-        let size = app_units::Au::from_f32_px(text_layout::px_to_pt(font_size));
-        if !self.font_instances.contains_key(&(name.to_owned(), size)) {
-            let instance_key = self.render.as_ref().unwrap().generate_font_instance_key();
-            let mut resources = ResourceUpdates::new();
-            resources.add_font_instance(instance_key, font_key, size, None, None, Vec::new());
-            self.render.as_ref().unwrap().update_resources(resources);
-            self.font_instances.insert((name.to_owned(), size), instance_key);
-        }
-        &self.font_instances[&(name.to_owned(), size)]
+    fn set_render_api(&mut self, render: RenderApiSender) {
+        self.render = Some(render.create_api());
+        self.font_loader.render = Some(render.create_api());
     }
 }
 fn load_image(file: &str) -> Result<(ImageData, ImageDescriptor), image::ImageError> {
@@ -238,22 +186,4 @@ pub fn premultiply(data: &mut [u8]) {
         pixel[1] = ((g * a + 128) / 255) as u8;
         pixel[0] = ((b * a + 128) / 255) as u8;
     }
-}
-
-fn load_system_font_by_family_name(name: &str) -> Result<Vec<u8>, ::std::io::Error> {
-    let property = system_fonts::FontPropertyBuilder::new().family(name).build();
-    let font = system_fonts::get(&property)
-        .map(|tuple| tuple.0)
-        .ok_or(::std::io::Error::new(::std::io::ErrorKind::NotFound, "Font not found"));
-    font
-}
-
-fn font_from_bytes(bytes: Vec<u8>) -> Result<Font, ::std::io::Error> {
-    let collection = rusttype::FontCollection::from_bytes(bytes);
-    let mut font_iter = collection.into_fonts();
-    font_iter.next().ok_or(::std::io::Error::new(::std::io::ErrorKind::InvalidData, "Bad font format"))
-}
-
-pub fn load_font(name: &str) -> Result<Font, ::std::io::Error> {
-    font_from_bytes(try!(load_system_font_by_family_name(name)))
 }
